[WIP] Prepare to switch MULTICLIENT003 to case statement analysis
